getCompressedBreakdownWithCount in typescript
